requirements.txt
# Company Research Assistant - Dependencies
# Install with: pip install -r requirements.txt

# Groq SDK for LLaMA-3.1-70B API access
groq>=0.4.0

# HTTP requests for Wikipedia API
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0

# Fuzzy string matching for company name correction
rapidfuzz>=3.0.0
#google-generativeai>=0.3.0

# Optional: Enhanced CLI experience (uncomment if desired)
# rich>=13.0.0
# prompt-toolkit>=3.0.0